+++ app/controllers/search_results_controller.rb
@@ -1,4 +1,5 @@
 class SearchResultsController < ApplicationController
+  include SearchResultsHelper
   before_action :set_search_result, only: %i[show edit update destroy]
 
   # GET /search_results
@@ -54,29 +55,6 @@ class SearchResultsController < ApplicationController
     people = Person.where("LOWER(first_name) || ' ' || LOWER(last_name) LIKE ?OR LOWER(last_name) LIKE ?",
                           "#{query}%", "#{query}%")
     add_search_results(rooms, buildings, locations, people)
-  end
-
-  # Helper functions for including querys to match the max linelengths
-  def including_buildings(query)
-    Building.where("LOWER(name) LIKE ? AND NOT LOWER(name) LIKE ? OR LOWER(name_de) LIKE ? AND NOT
-      LOWER(name_de) LIKE ? ", "%#{query}%", "#{query}%", "%#{query}%", "#{query}%")
-  end
-
-  def including_rooms(query)
-    Room.where("LOWER(name) || LOWER(room_type) LIKE ? AND NOT LOWER(name) || LOWER(room_type) LIKE ? OR
-      LOWER(name_de) LIKE ? AND NOT LOWER(name_de) LIKE ? ", "%#{query}%", "#{query}%", "%#{query}%", "#{query}%")
-  end
-
-  def including_locations(query)
-    Location.where("LOWER(name) LIKE ? AND NOT LOWER(name) LIKE ? OR LOWER(name_de) LIKE ? AND NOT
-    LOWER(name_de) LIKE ? OR LOWER(details_de) LIKE ? AND NOT LOWER(details_de) LIKE ? OR LOWER(details) LIKE ?
-    AND NOT LOWER(details) LIKE ?", "%#{query}%", "#{query}%", "%#{query}%", "#{query}%", "%#{query}%", "#{query}%",
-                   "%#{query}%", "#{query}%")
-  end
-
-  def including_people(query)
-    Person.where("LOWER(first_name) || ' ' || LOWER(last_name) LIKE ? AND NOT LOWER(first_name) || ' ' ||
-      LOWER(last_name) LIKE ? AND NOT LOWER(last_name) LIKE ?", "%#{query}%", "#{query}%", "#{query}%")
   end
 
   def search_for_entries_including(query)
